requirements.txt
asgiref==3.8.1
attrs==25.4.0
Django==5.1.3
django-modeltranslation==0.19.7
django-phonenumber-field==8.4.0
djangorestframework==3.15.2
dotenv==0.9.9
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==24.2
phonenumbers==9.0.19
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.2
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
